some changes in windws container
